better props cast
