+++ android/src/main/java/com/callstack/reactnativebrownfield/ReactNativeFragment.kt
@@ -1,8 +1,6 @@
 package com.callstack.reactnativebrownfield;
 
 import android.annotation.TargetApi
-import android.content.Context
-import android.content.Intent
 import android.os.Build
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -13,8 +11,6 @@ import android.view.ViewGroup
 import com.facebook.infer.annotation.Assertions
 import com.facebook.react.ReactRootView
 import com.facebook.react.bridge.Callback
-import com.facebook.react.bridge.ReadableArray
-import com.facebook.react.bridge.ReadableMap
 import com.facebook.react.bridge.WritableMap
 import com.facebook.react.common.LifecycleState
 import com.facebook.react.devsupport.DoubleTapReloadRecognizer
@@ -28,14 +24,13 @@ private const val INITIAL_PROPS = "com.callstack.reactnativebrownfield.FRAGMENT_
 class ReactNativeFragment : Fragment(), PermissionAwareActivity {
 
     private var reactRootView: ReactRootView? = null
-    private lateinit var moduleName: String
     private lateinit var doubleTapReloadRecognizer: DoubleTapReloadRecognizer
     private lateinit var permissionsCallback: Callback
     private var permissionListener: PermissionListener? = null
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        moduleName = arguments?.getString(MODULE_NAME)!!
+        val moduleName = arguments?.getString(MODULE_NAME)!!
         val initialProps = arguments?.getBundle(INITIAL_PROPS)
 
         doubleTapReloadRecognizer = DoubleTapReloadRecognizer()
@@ -44,7 +39,7 @@ class ReactNativeFragment : Fragment(), PermissionAwareActivity {
         reactRootView?.startReactApplication(
             BridgeManager.shared.reactNativeHost.reactInstanceManager,
             moduleName,
-            null
+            initialProps
         )
     }
 
@@ -167,36 +162,13 @@ class ReactNativeFragment : Fragment(), PermissionAwareActivity {
         }
 
         @JvmStatic
-        fun createReactNativeFragment(moduleName: String, initialProps: WritableMap): ReactNativeFragment {
-            val bundle = Bundle()
-            initialProps.toHashMap().forEach {
-                when (it.value) {
-                    // Should we go recursively with arrays and maps?
-                    is ReadableArray -> {
-                        bundle.putSerializable(it.key, (it.value as ReadableArray).toArrayList())
-                    }
-                    is ReadableMap -> {
-                        bundle.putSerializable(it.key, (it.value as ReadableMap).toHashMap())
-                    }
-                    is Boolean -> {
-                        bundle.putBoolean(it.key, it.value as Boolean)
-                    }
-                    is Int -> {
-                        bundle.putInt(it.key, it.value as Int)
-                    }
-                    is String -> {
-                        bundle.putString(it.key, it.value as String)
-                    }
-                    is Double -> {
-                        bundle.putDouble(it.key, it.value as Double)
-                    }
-                    else -> {
-                        bundle.putSerializable(it.key, null)
-                    }
-                }
-            }
+        fun createReactNativeFragment(moduleName: String, initialProps: HashMap<String, *>): ReactNativeFragment {
+            return createReactNativeFragment(moduleName, PropsBundle.fromHashMap(initialProps))
+        }
 
-            return createReactNativeFragment(moduleName, bundle)
+        @JvmStatic
+        fun createReactNativeFragment(moduleName: String, initialProps: WritableMap): ReactNativeFragment {
+            return createReactNativeFragment(moduleName, initialProps.toHashMap())
         }
     }
